Add interface to search class, refactor search class
