+++ DualWallpaper/Interfaces/ISearch.cs
@@ -1,0 +1,9 @@
+﻿using System.Windows.Forms;
+
+namespace DualWallpaper.Interfaces
+{
+    public interface ISearch
+    {
+        void OpenDefaultBrowserWithWallpaperSearchResults(Panel panel);
+    }
+}
